separates merging utility functions into their own file (#1629)
Adds a merge_utils.py file that contains all of the functionality to merge datasets with numbers from big query.
